requirements.txt
fastapi==0.111.0
google-generativeai==0.7.2
openai==1.30.1
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.7.1